Added first pass of camera functionality
Added functions for restricting the drawing of sprites and drawing sprites relative to the position of a camera that can exist in a Uint64*Uint64 size world. There are still a few bugs with Uint64 overflow but at the moment, main.c should draw sprites to a 100*100 camera centred on entity 1.
Functionality will soon be used to add splitscreen coop.